Move serialization logic to its own module
