+++ src/renderer/components/TranscriptEditor.tsx
@@ -1,6 +1,10 @@
-import { userHasChosenTranscriptFile } from "../../common/ipcChannelNames"
+import {
+  heresTheTranscript,
+  userHasChosenTranscriptFile,
+} from "../../common/ipcChannelNames"
 import { ipcRenderer } from "electron"
 import React, { useMemo, useState, useEffect } from "react"
+import { deserializeTranscript, serializeTranscript } from "../serialization"
 import { createEditor, Node } from "slate"
 import { Slate, Editable, withReact } from "slate-react"
 
@@ -11,17 +15,6 @@ const initialTranscriptValue: Node[] = [
   },
 ]
 
-function deserializeTranscript(transcript: string): Node[] {
-  const value = transcript.split("\n").map((s) => {
-    return {
-      type: "paragraph",
-      children: [{ text: s }],
-    }
-  })
-  console.log("transcript is", value)
-  return value
-}
-
 function App(): JSX.Element {
   const editor = useMemo(() => withReact(createEditor()), [])
   const [value, setValue] = useState<Node[]>(initialTranscriptValue)
@@ -31,6 +24,7 @@ function App(): JSX.Element {
     ipcRenderer.on(userHasChosenTranscriptFile, (event, transcript) =>
       setValue(deserializeTranscript(transcript)),
     )
+
     return function cleanup(): void {
       ipcRenderer.removeListener(
         userHasChosenTranscriptFile,
@@ -46,7 +40,10 @@ function App(): JSX.Element {
       id="editor"
       editor={editor}
       value={value}
-      onChange={(value): void => setValue(value)}
+      onChange={(value): void => {
+        setValue(value)
+        ipcRenderer.send(heresTheTranscript, serializeTranscript(value))
+      }}
     >
       <Editable placeholder="Drag a text file here to open it, or just start typing!" />
     </Slate>
